Some refactoring of common code
